Shop Price Ranomizer + Correct Item Collect Test
First test of displaying the correct item text when collecting an item. Current issues seen so far are phobekin portraits being off position and shop collections still show original item text.

Shop prices have also been randomized if toggled in the menu

Toggles are not saved and can cause issues if reloading a save with them in states they were not generated with.